Adds Ranking System for PHM Flight Modes
A function is added to help determine if the vehicle is already in
more sever flight mode than the one that the fault incidated by
the PHM device intends to set. For example, a warning fault will
not set the mode to STABILIZE if the vehicle is in LAND mode.